[FEATURE] Add support for (most) WooCommerce settings.
Augments the existing site and network states rather than adds new
states as it doesn't really make sense as a state.

Due to single level of regions it feels a little clunky, in that you
need to specify each option at the same level, e.g.

`woocommerce-general`

as opposed to

```yaml
woocommerce:
  general:
```

There needs to be a bit of work in the dictator library to better
support nested regions and inherited regions and stuff, but for now this
will work.